Move StandardErrorCurve to its own module, add named constructors
The curve was defined in experiment_analysis but needed by
relative_lift_transformer, which imports the other way round. That forced a
deferred import inside standard_error_curve() to dodge the cycle. Its own module
removes the cycle, so the transformer imports it at the top like anything else
and the method gets a real return annotation.

Also makes construction explicit at the call sites rather than implied by which
defaults are omitted:

- StandardErrorCurve.constant(se) says the standard error does not vary with the
  effect, instead of leaving that to be inferred from two absent arguments.
- StandardErrorCurve.from_independent_arms(se, effect_var) encodes that
  independent arms force effect_cov == -effect_var, so callers cannot get that
  sign wrong. The delta-method transformer had been setting it by hand.

is_effect_dependent stays derived. A stored flag would default to a value that
can contradict the coefficients it describes, so it would need validating against
them, which is strictly worse than computing it from them.

Restores the multi-line f-string in ExperimentAnalysis.__str__ that ruff format
had joined. Note ruff.toml sets line-length 160 while pre-commit runs black at
its default 88, so the two formatters disagree on anything in between; joining an
implicit concatenation is not reversible by black, so it sticks.
